agents/requirements.txt

# OpenAI API for GPT-4
openai>=1.9.0
requests>=2.28.0
python-dotenv>=1.0.0
pytest>=7.0.0
